feat: :art: Moved interfaces to a new directory
IRemoved interfaces from domManipulation, created a directory types and added interfaces to a new file called customTypes
