+++ src/types/customTypes.d.ts
@@ -1,0 +1,12 @@
+interface Temperature {
+    max: string,
+    min: string
+  }
+
+  export interface Forecast{
+  
+    date: string, 
+    weather: string,
+    temp2m: Temperature,
+  
+}
